before core business

+++ src/SchoolProject.UserModule.Api/Controllers/UserController.cs
@@ -1,9 +1,11 @@
 using AutoMapper;
 using Microsoft.AspNetCore.Mvc;
 using SchoolProject.UserModule.Api.DTOs;
+using SchoolProject.UserModule.Api.ExceptionHandler;
 using SchoolProject.UserModule.Api.Filter;
 using SchoolProject.UserModule.Business.Models;
 using SchoolProject.UserModule.Business.Repositories.Interfaces;
+using SchoolProject.UserModule.Business.Services.Interfaces;
 
 namespace SchoolProject.UserModule.Api.Controllers
 {
@@ -14,11 +16,13 @@ namespace SchoolProject.UserModule.Api.Controllers
     {
         private readonly IUserRepo _userRepo;
         private readonly IMapper _mapper;
+        private readonly IAuthService _authService;
 
-        public UserController(IUserRepo userRepo, IMapper mapper)
+        public UserController(IUserRepo userRepo, IMapper mapper, IAuthService authService)
         {
             _userRepo = userRepo;
             _mapper = mapper;
+            _authService = authService;
         }
 
         [HttpGet]
@@ -33,39 +37,43 @@ namespace SchoolProject.UserModule.Api.Controllers
         public async Task<IActionResult> GetUserById(int id)
         {
             var user = await _userRepo.GetUserById(id);
-            var userDto = _mapper.Map<GetUserDto>(user);
-            return Ok(userDto);
+            if (user != null)
+            {
+                var userDto = _mapper.Map<GetUserDto>(user);
+                return Ok(userDto);
+            }
+            return NotFound(ErrorMsgConstant.UserNotFound);
         }
 
         [HttpPost]
-        public async Task<IActionResult> AddUser([FromBody] PostUserDto userDto)
+        public async Task<IActionResult> AddUser(PostUserDto userDto)
         {
+            userDto.Password = _authService.HashPasswordWithSalt(userDto.Password);
             User user = _mapper.Map<User>(userDto);
-            try
+            User newUser = await _userRepo.AddUser(user);
+            if (newUser != null)
             {
-                User newUser = await _userRepo.AddUser(user);
                 var newUserDto = _mapper.Map<GetUserDto>(newUser);
                 return Ok(newUserDto);
             }
-            catch (Exception ex)
-            {
-                throw new Exception("jai shree ram" + ex.Message);   
-            }
+            return BadRequest(ErrorMsgConstant.UserNotCreated);
         }
 
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteUser(int id)
         {
-            try
+            var user = await _userRepo.GetUserById(id);
+            if (user != null)
             {
-                var user = await _userRepo.GetUserById(id);
-                await _userRepo.DeleteUser(user);
-                return Ok();
+                bool success = await _userRepo.DeleteUser(user);
+                if (success)
+                {
+                    var userDto = _mapper.Map<GetUserDto>(user);
+                    return Ok(userDto);
+                }
+                throw new Exception(ErrorMsgConstant.UserAlreadyDeleted);
             }
-            catch (Exception ex)
-            {
-                return BadRequest(new { message = ex.Message });
-            }
+            return NotFound(ErrorMsgConstant.UserNotFound);
         }
     }
 }
